requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
certifi==2024.2.2
click==8.1.7
distro==1.9.0
docstring_parser==0.16
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
instructor==1.2.2
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
openai==1.23.2
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
python-dotenv==1.0.1
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
tenacity==8.2.3
tk==0.1.0
tqdm==4.66.2
typer==0.12.3
typing_extensions==4.11.0
yarl==1.9.4
